Cleanup and Modification
Motor instructions dont need to continuously call their setPower() methods as setting it once will keep it at that. This defeats the purpose of a duration attribute for the motor instruction class. Moving instruction automatically call the StopDriving() method afterwards so as to reduce the necessity of adding an instruction to stop after an instruction to move.